requirements.txt
flask==3.0.0
plivo==4.47.0
python-dotenv==1.0.0
